+++ assets/javascripts/kitten/components/form/radio-button/index.js
@@ -231,10 +231,42 @@ const StyledRadioButton = styled.div`
       }
     }
 
-    .k-Form-RadioButton__labelText {
-      ${TYPOGRAPHY.fontStyles.regular};
-      line-height: ${pxToRem(20)};
-      font-size: ${stepToRem(-1)};
+    /* PARAGRAPH STYLE */
+
+    &.k-Form-RadioButton--paragraphStyle {
+      .k-Form-RadioButton__label {
+        align-items: center;
+      }
+
+      .k-Form-RadioButton__label::before {
+        margin-right: ${pxToRem(15)};
+      }
+
+      .k-Form-RadioButton__labelText {
+        flex: 1 0 calc(100% - ${pxToRem(20 + 15)});
+      }
+    }
+
+    /* FONT STYLES */
+
+    &.k-Form-RadioButton__labelText--regular {
+      .k-Form-RadioButton__labelText {
+        ${TYPOGRAPHY.fontStyles.regular};
+        line-height: ${pxToRem(20)};
+        font-size: ${stepToRem(-1)};
+      }
+    }
+
+    &.k-Form-RadioButton__labelText--light {
+      .k-Form-RadioButton__labelText {
+        ${TYPOGRAPHY.fontStyles.light};
+      }
+    }
+
+    &.k-Form-RadioButton__labelText--bold {
+      .k-Form-RadioButton__labelText {
+        ${TYPOGRAPHY.fontStyles.bold};
+      }
     }
 
     &.k-Form-RadioButton--small {
@@ -329,6 +361,8 @@ export const RadioButton = ({
   disabled,
   variant,
   design,
+  fontWeight,
+  paragraphStyle,
   ...inputProps
 }) => {
   return (
@@ -339,10 +373,12 @@ export const RadioButton = ({
         `k-Form-RadioButton--${variant}`,
         `k-Form-RadioButton--${design}`,
         `k-Form-RadioButton--${size}`,
+        `k-Form-RadioButton__labelText--${fontWeight}`,
         {
           'k-Form-RadioButton--error': error,
           'k-Form-RadioButton--largeLabel': large,
           'k-Form-RadioButton--largeContent': largeContent,
+          'k-Form-RadioButton--paragraphStyle': paragraphStyle,
         },
       )}
     >
@@ -380,6 +416,8 @@ RadioButton.propTypes = {
   variant: PropTypes.oneOf(['andromeda', 'orion']),
   design: PropTypes.oneOf(['disc', 'check']),
   size: PropTypes.oneOf(['small', 'regular', 'big']),
+  fontWeight: PropTypes.oneOf(['light', 'regular', 'bold']),
+  paragraphStyle: PropTypes.bool,
 }
 
 RadioButton.defaultProps = {
@@ -390,4 +428,6 @@ RadioButton.defaultProps = {
   variant: 'orion',
   design: 'disc',
   size: 'regular',
+  fontWeight: 'regular',
+  paragraphStyle: false,
 }